last-1.0.0.0

+++ src/Comp/Users.jsx
@@ -1,17 +1,38 @@
 import React, { useState } from 'react';
 import { useLoaderData } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const Users = () => {
     const data = useLoaderData();
-    const [us,setUs] = useState(data);
+    const [us,setUs] = useState(data); // us = user 
     const handleDelete = (e) =>{
-        fetch(`http://localhost:5100/user/${e}`,{
+      Swal.fire({
+  title: "Are you sure?",
+  text: "You won't be able to revert this!",
+  icon: "warning",
+  showCancelButton: true,
+  confirmButtonColor: "#3085d6",
+  cancelButtonColor: "#d33",
+  confirmButtonText: "Yes, delete it!"
+}).then((result) => {
+  if (result.isConfirmed) {
+    fetch(`http://localhost:5100/user/${e}`,{
             method: 'DELETE'
         })
         .then(res=>res.json())
         .then(info=>{
+          const remain = us.filter(infos=>infos._id!==e)
+          setUs(remain);
             console.log(info);
         })
+    Swal.fire({
+      title: "Deleted!",
+      text: "Your file has been deleted.",
+      icon: "success"
+    });
+  }
+});
+        
     }
     return (
         <div>
